transform all int/logical types to double for getParam and getBound; added tests for integer-valued user-defined dist, including with integer-valued parameter

+++ packages/nimble/inst/tests/test-getParam.R
@@ -137,7 +137,7 @@ test_that('getParam, three-dimensional', {
 test_that('getParam, user-defined integer-valued', {
     dtest <- nimbleFunction(
         run = function(x = integer(0), thetaInt = integer(0), thetaDbl = double(0), log = integer(0, default = 0)) {
-            returnType(integer(0))
+            returnType(double(0))
             return(0)
         })
     temporarilyAssignInGlobalEnv(dtest)
